back end nuevos cambios/requirements.txt
asgiref==3.8.1
bcrypt==5.0.0
breadability==0.1.20
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
chilean-rut==0.1.3
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-database-url==3.0.1
Django==5.1.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.7.0
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.55.0
gunicorn==23.0.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
nltk==3.9.1
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.4
packaging==24.2
Panda3D==1.10.14
panda3d-gltf==1.2.0
panda3d-simplepbr==0.12.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.11
pycountry==24.6.1
pymongo==4.13.2
PyMySQL==1.1.2
pyparsing==3.2.0
pyperclip==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rutificador==0.2.17
rutpy==1.0.2
screeninfo==0.8.1
six==1.16.0
sqlparse==0.5.1
sumy==0.11.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
ursina==7.0.0
whitenoise==6.11.0
